+++ windows/flutter/generated_plugins.cmake
@@ -4,6 +4,7 @@
 
 list(APPEND FLUTTER_PLUGIN_LIST
   auto_updater_windows
+  desktop_drop
   url_launcher_windows
 )